+++ src/pages/Home.jsx
@@ -1,0 +1,107 @@
+import Frame from "../components/Frame";
+import Button from "../components/Button";
+import Panel from "../components/Panel";
+import Servers from "./Servers";
+import { useState } from "react";
+import { useNavigate } from 'react-router-dom';
+
+export default function Home(props) {
+  const [serversVisible, setServersVisible] = useState(false);
+
+  function createServers() {
+    return(
+      <Servers />
+    )
+  }
+
+  function handleClick(){
+    serversVisible ? setServersVisible(false) : serversVisible(true);
+  }
+
+  return (
+    <>
+      <Frame>
+        <div className="flex justify-center -mt-24">
+          <div className="flex flex-col justify-center h-screen">
+            <Panel>
+              <h1 className="text-7xl font-header text-green-50">jake</h1>
+              <h1 className="text-4xl -mt-2 font-header text-green-50">
+                dubeau
+              </h1>
+              <div className="flex gap-3 justify-center">
+                <div className="my-2">
+                  <Button text="GitHub" to="https://github.com/StinkyDubeau">
+                    <div className="flex justify-center gap-2 fill-green-50">
+                      <svg
+                        xmlns="http://www.w3.org/2000/svg"
+                        width="16"
+                        height="16"
+                        className="bi bi-github"
+                        viewBox="0 0 16 16"
+                      >
+                        <path d="M8 0C3.58 0 0 3.58 0 8c0 3.54 2.29 6.53 5.47 7.59.4.07.55-.17.55-.38 0-.19-.01-.82-.01-1.49-2.01.37-2.53-.49-2.69-.94-.09-.23-.48-.94-.82-1.13-.28-.15-.68-.52-.01-.53.63-.01 1.08.58 1.23.82.72 1.21 1.87.87 2.33.66.07-.52.28-.87.51-1.07-1.78-.2-3.64-.89-3.64-3.95 0-.87.31-1.59.82-2.15-.08-.2-.36-1.02.08-2.12 0 0 .67-.21 2.2.82.64-.18 1.32-.27 2-.27s1.36.09 2 .27c1.53-1.04 2.2-.82 2.2-.82.44 1.1.16 1.92.08 2.12.51.56.82 1.27.82 2.15 0 3.07-1.87 3.75-3.65 3.95.29.25.54.73.54 1.48 0 1.07-.01 1.93-.01 2.2 0 .21.15.46.55.38A8.01 8.01 0 0 0 16 8c0-4.42-3.58-8-8-8" />
+                      </svg>
+                    </div>
+                  </Button>
+                </div>{" "}
+                <div className="my-2">
+                  <Button
+                    text="LinkedIn"
+                    to="https://www.linkedin.com/in/dubeau-jake"
+                  >
+                    <div className="flex justify-center gap-2 fill-green-50">
+                      <svg
+                        xmlns="http://www.w3.org/2000/svg"
+                        width="16"
+                        height="16"
+                        fill="currentColor"
+                        className="bi bi-linkedin fill-green-50"
+                        viewBox="0 0 16 16"
+                      >
+                        <path d="M0 1.146C0 .513.526 0 1.175 0h13.65C15.474 0 16 .513 16 1.146v13.708c0 .633-.526 1.146-1.175 1.146H1.175C.526 16 0 15.487 0 14.854zm4.943 12.248V6.169H2.542v7.225zm-1.2-8.212c.837 0 1.358-.554 1.358-1.248-.015-.709-.52-1.248-1.342-1.248S2.4 3.226 2.4 3.934c0 .694.521 1.248 1.327 1.248zm4.908 8.212V9.359c0-.216.016-.432.08-.586.173-.431.568-.878 1.232-.878.869 0 1.216.662 1.216 1.634v3.865h2.401V9.25c0-2.22-1.184-3.252-2.764-3.252-1.274 0-1.845.7-2.165 1.193v.025h-.016l.016-.025V6.169h-2.4c.03.678 0 7.225 0 7.225z" />
+                      </svg>
+                    </div>
+                  </Button>
+                </div>
+                <div className="my-2"  onClick={useNavigate("/servers")}>
+                  <Button text="Servers">
+                    <div className="flex justify-center gap-2 fill-green-50">
+                      <svg
+                        xmlns="http://www.w3.org/2000/svg"
+                        width="16"
+                        height="16"
+                        fill="currentColor"
+                        className="bi bi-dpad-fill fill-green-50"
+                        viewBox="0 0 16 16"
+                      >
+                        <path d="M6.5 0A1.5 1.5 0 0 0 5 1.5v3a.5.5 0 0 1-.5.5h-3A1.5 1.5 0 0 0 0 6.5v3A1.5 1.5 0 0 0 1.5 11h3a.5.5 0 0 1 .5.5v3A1.5 1.5 0 0 0 6.5 16h3a1.5 1.5 0 0 0 1.5-1.5v-3a.5.5 0 0 1 .5-.5h3A1.5 1.5 0 0 0 16 9.5v-3A1.5 1.5 0 0 0 14.5 5h-3a.5.5 0 0 1-.5-.5v-3A1.5 1.5 0 0 0 9.5 0zm1.288 2.34a.25.25 0 0 1 .424 0l.799 1.278A.25.25 0 0 1 8.799 4H7.201a.25.25 0 0 1-.212-.382zm0 11.32-.799-1.277A.25.25 0 0 1 7.201 12H8.8a.25.25 0 0 1 .212.383l-.799 1.278a.25.25 0 0 1-.424 0Zm-4.17-4.65-1.279-.798a.25.25 0 0 1 0-.424l1.279-.799A.25.25 0 0 1 4 7.201V8.8a.25.25 0 0 1-.382.212Zm10.043-.798-1.278.799A.25.25 0 0 1 12 8.799V7.2a.25.25 0 0 1 .383-.212l1.278.799a.25.25 0 0 1 0 .424Z" />
+                      </svg>
+                    </div>
+                  </Button>
+                </div>
+                {/* <div className="my-2">
+                <Button text="Astros" to="">
+                <a className="flex justify-center gap-2 fill-green-50">
+                <svg
+                xmlns="http://www.w3.org/2000/svg"
+                width="16"
+                height="16"
+                fill="currentColor"
+                className="bi bi-camera-fill fill-green-50"
+                      viewBox="0 0 16 16"
+                      >
+                      <path d="M10.5 8.5a2.5 2.5 0 1 1-5 0 2.5 2.5 0 0 1 5 0" />
+                      <path d="M2 4a2 2 0 0 0-2 2v6a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V6a2 2 0 0 0-2-2h-1.172a2 2 0 0 1-1.414-.586l-.828-.828A2 2 0 0 0 9.172 2H6.828a2 2 0 0 0-1.414.586l-.828.828A2 2 0 0 1 3.172 4zm.5 2a.5.5 0 1 1 0-1 .5.5 0 0 1 0 1m9 2.5a3.5 3.5 0 1 1-7 0 3.5 3.5 0 0 1 7 0" />
+                      </svg>
+                      </a>
+                      </Button>
+                    </div> */}
+              </div>
+            </Panel>
+            {serversVisible && createServers()}
+          </div>
+        </div>
+      </Frame>
+    </>
+  );
+}
